phdkim_main_capture/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
pycparser==2.22
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.1.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
